+++ packages/brick_sqlite_generators/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Unreleased
 
+## 2.0.3
+
+* Fix serializing iterable `toJson` classes to SQLite
+* Fix migration generation for iterable `toJson` classes
+
 ## 2.0.2
 
 * Loosen dependency restrictions to major versions